backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
cryptography
fastapi
uvicorn
pydantic
python-dotenv
cryptography
